src/requirements.txt
action-msgs==2.0.2
action-tutorials-interfaces==0.33.5
action-tutorials-py==0.33.5
actionlib-msgs==5.3.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
ament-cmake-test==2.5.4
ament-copyright==0.17.2
ament-cppcheck==0.17.2
ament-cpplint==0.17.2
ament-flake8==0.17.2
ament-index-python==1.8.1
ament-lint==0.17.2
ament-lint-cmake==0.17.2
ament-package==0.16.4
ament-pep257==0.17.2
ament-uncrustify==0.17.2
ament-xmllint==0.17.2
angles==1.16.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
builtin-interfaces==2.0.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.1
coloredlogs==15.0.1
composition-interfaces==2.0.2
cv-bridge==4.1.0
dataclasses-json==0.6.7
demo-nodes-py==0.33.5
diagnostic-msgs==5.3.6
distro==1.9.0
domain-coordinator==0.12.0
durationpy==0.10
example-interfaces==0.12.0
examples-rclpy-executors==0.19.5
examples-rclpy-minimal-action-client==0.19.5
examples-rclpy-minimal-action-server==0.19.5
examples-rclpy-minimal-client==0.19.5
examples-rclpy-minimal-publisher==0.19.5
examples-rclpy-minimal-service==0.19.5
examples-rclpy-minimal-subscriber==0.19.5
fastapi==0.123.5
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
geometry-msgs==5.3.6
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
image-geometry==4.1.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
interactive-markers==2.5.4
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.1.0
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-google-genai==3.2.0
langchain-huggingface==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langsmith==0.4.53
laser-geometry==2.7.1
launch==3.4.5
launch-ros==0.26.8
launch-testing==3.4.5
launch-testing-ros==0.26.8
launch-xml==3.4.5
launch-yaml==3.4.5
lifecycle-msgs==2.0.2
logging-demo==0.33.5
map-msgs==2.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
message-filters==4.11.6
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nav-msgs==5.3.6
networkx==3.6
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
osrf-pycommon==2.1.6
overrides==7.7.0
packaging==25.0
pcl-msgs==1.0.0
pendulum-msgs==0.33.5
pillow==12.0.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-qt-binding==2.2.1
PyYAML==6.0.3
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.5
rcl-interfaces==2.0.2
rclpy==7.1.4
rcutils==6.7.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resource-retriever==3.4.4
rich==14.2.0
rmw-dds-common==3.1.0
ros2action==0.32.4
ros2bag==0.26.7
ros2bag-mcap-cli==0.26.7
ros2bag-sqlite3-cli==0.26.7
ros2cli==0.32.4
ros2component==0.32.4
ros2doctor==0.32.4
ros2interface==0.32.4
ros2launch==0.26.8
ros2lifecycle==0.32.4
ros2multicast==0.32.4
ros2node==0.32.4
ros2param==0.32.4
ros2pkg==0.32.4
ros2run==0.32.4
ros2service==0.32.4
ros2topic==0.32.4
rosbag2-interfaces==0.26.7
rosbag2-py==0.26.7
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.5
rosidl-cli==4.6.5
rosidl-cmake==4.6.5
rosidl-generator-c==4.6.5
rosidl-generator-cpp==4.6.5
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.5
rosidl-parser==4.6.5
rosidl-pycommon==4.6.5
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.1
rosidl-typesupport-fastrtps-cpp==3.6.1
rosidl-typesupport-introspection-c==4.6.5
rosidl-typesupport-introspection-cpp==4.6.5
rpds-py==0.30.0
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.4
rqt-bag-plugins==1.5.4
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.3
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
sentence-transformers==5.1.2
service-msgs==2.0.2
setuptools==80.9.0
shape-msgs==5.3.6
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.44
sros2==0.13.3
starlette==0.50.0
statistics-msgs==2.0.2
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tf2-geometry-msgs==0.36.12
tf2-kdl==0.36.12
tf2-msgs==0.36.12
tf2-py==0.36.12
tf2-ros-py==0.36.12
tf2-sensor-msgs==0.36.12
tf2-tools==0.36.12
threadpoolctl==3.6.0
tokenizers==0.22.1
topic-monitor==0.33.5
torch==2.9.1
tqdm==4.67.1
trajectory-msgs==5.3.6
transformers==4.57.3
triton==3.5.1
turtlesim==1.8.3
type-description-interfaces==2.0.2
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.5.0
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
visualization-msgs==5.3.6
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
